Move generic dependency machinery into Graphula.Internal
